refactor: extract quick-pick type-to-filter helper and fix constants filename typo
- add selectQuickInputOptionByTyping to playwright-vscode-ext, replacing the
  duplicated quick-pick + filter + click sequence in the colorizer spec
- rename contants.ts -> constants.ts and update all 8 importing specs
