+++ packages/salesforcedx-vscode-apex-testing/test/playwright/specs/runApexTestsFailAndFix.headless.spec.ts
@@ -32,7 +32,7 @@ import {
 
 import packageNls from '../../../package.nls.json';
 import { test } from '../fixtures';
-import { TEST_RUN_TIMEOUT } from '../contants';
+import { TEST_RUN_TIMEOUT } from '../constants';
 import { CMD_TOGGLE_MAXIMIZED_PANEL } from '../helpers/testExplorerHelpers';
 
 const ACCOUNT_SERVICE_CONTENT = [
